Create DefaultCredentialsProvider
Fixes the missing bean issue in the config sample.
