+++ spring-cloud-gcp-autoconfigure/src/main/java/org/springframework/cloud/gcp/autoconfigure/core/GcpContextAutoConfiguration.java
@@ -16,54 +16,30 @@
 
 package org.springframework.cloud.gcp.autoconfigure.core;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 import com.google.api.gax.core.CredentialsProvider;
-import com.google.api.gax.core.FixedCredentialsProvider;
-import com.google.api.gax.core.GoogleCredentialsProvider;
-import com.google.auth.oauth2.ComputeEngineCredentials;
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.auth.oauth2.ServiceAccountCredentials;
-import com.google.auth.oauth2.UserCredentials;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.gcp.core.Credentials;
+import org.springframework.cloud.gcp.core.DefaultCredentialsProvider;
 import org.springframework.cloud.gcp.core.DefaultGcpProjectIdProvider;
 import org.springframework.cloud.gcp.core.GcpProjectIdProvider;
-import org.springframework.cloud.gcp.core.GcpScope;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.util.ObjectUtils;
-import org.springframework.util.StringUtils;
+
 
 /**
- * Base starter for Google Cloud Projects. Provides defaults for {@link GoogleCredentials}.
+ * Base starter for Google Cloud Projects. Provides defaults for {@link com.google.auth.oauth2.GoogleCredentials}.
  * Binds properties from {@link GcpProperties}.
  *
  * @author Vinicius Carvalho
  * @author João André Martins
+ * @author Mike Eltsufin
  */
 @Configuration
 @EnableConfigurationProperties(GcpProperties.class)
 public class GcpContextAutoConfiguration {
-	private static final String DEFAULT_SCOPES_PLACEHOLDER = "DEFAULT_SCOPES";
-
-	private static final List<String> CREDENTIALS_SCOPES_LIST =
-			Collections.unmodifiableList(
-					Arrays.stream(GcpScope.values())
-						.map(GcpScope::getUrl)
-						.collect(Collectors.toList()));
-
 	private static final Log LOGGER = LogFactory.getLog(GcpContextAutoConfiguration.class);
 
 	private final GcpProperties gcpProperties;
@@ -72,69 +48,11 @@ public class GcpContextAutoConfiguration {
 		this.gcpProperties = gcpProperties;
 	}
 
-	protected List<String> resolveScopes() {
-		Credentials propertyCredentials = this.gcpProperties.getCredentials();
-		if (!ObjectUtils.isEmpty(propertyCredentials.getScopes())) {
-			Set<String> resolvedScopes = new HashSet<>();
-			propertyCredentials.getScopes().forEach(scope -> {
-				if (DEFAULT_SCOPES_PLACEHOLDER.equals(scope)) {
-					resolvedScopes.addAll(CREDENTIALS_SCOPES_LIST);
-				}
-				else {
-					resolvedScopes.add(scope);
-				}
-			});
-
-			return Collections.unmodifiableList(new ArrayList<>(resolvedScopes));
-		}
-
-		return CREDENTIALS_SCOPES_LIST;
-	}
-
 	@Bean
 	@ConditionalOnMissingBean
 	public CredentialsProvider googleCredentials() throws Exception {
-		CredentialsProvider credentialsProvider;
-
-		Credentials propertyCredentials = this.gcpProperties.getCredentials();
-
-		List<String> scopes = resolveScopes();
-
-		if (!StringUtils.isEmpty(propertyCredentials.getLocation())) {
-			credentialsProvider = FixedCredentialsProvider
-					.create(GoogleCredentials.fromStream(
-							propertyCredentials.getLocation().getInputStream())
-					.createScoped(scopes));
-		}
-		else {
-			credentialsProvider = GoogleCredentialsProvider.newBuilder()
-					.setScopesToApply(scopes)
-					.build();
-		}
-
-		try {
-			com.google.auth.Credentials credentials = credentialsProvider.getCredentials();
-
-			if (LOGGER.isInfoEnabled()) {
-				if (credentials instanceof UserCredentials) {
-					LOGGER.info("Default credentials provider for user "
-							+ ((UserCredentials) credentials).getClientId());
-				}
-				else if (credentials instanceof ServiceAccountCredentials) {
-					LOGGER.info("Default credentials provider for service account "
-							+ ((ServiceAccountCredentials) credentials).getClientEmail());
-				}
-				else if (credentials instanceof ComputeEngineCredentials) {
-					LOGGER.info("Default credentials provider for Google Compute Engine.");
-				}
-				LOGGER.info("Scopes in use by default credentials: " + scopes.toString());
-			}
-		}
-		catch (IOException ioe) {
-			LOGGER.error("No credentials were found.", ioe);
-		}
-
-		return credentialsProvider;
+		return new DefaultCredentialsProvider(this.gcpProperties.getCredentials().getScopes(),
+				this.gcpProperties.getCredentials().getLocation());
 	}
 
 	/**
